SP/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==1.0
requests>=2.31.0
python-dotenv==0.19.1
nltk==3.6.5
joblib==1.1.0
yfinance>=0.2.31
backoff>=2.2.1